requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
bcrypt==5.0.0
python-jose[cryptography]==46.0.7